feat(cosmos)!: compile agd binary to supercede ag-chain-cosmos
