+++ golang/cosmos/cmd/helper/main.go
@@ -6,6 +6,7 @@ import (
 
 	gaia "github.com/Agoric/agoric-sdk/golang/cosmos/app"
 	"github.com/Agoric/agoric-sdk/golang/cosmos/daemon"
+	daemoncmd "github.com/Agoric/agoric-sdk/golang/cosmos/daemon/cmd"
 )
 
 func main() {
@@ -15,5 +16,6 @@ func main() {
 	}
 
 	gaia.DefaultNodeHome = filepath.Join(userHomeDir, ".ag-cosmos-helper")
+	daemoncmd.AppName = "ag-cosmos-helper"
 	daemon.Run()
 }
